software/requirements.txt
RPi.GPIO==0.7.1
requests==2.31.0
paho-mqtt==2.0.0
python-dotenv==1.0.1
